+++ Api-Tools-Demo/Api-ToolsTests/TodoFakeModel.swift
@@ -1,0 +1,13 @@
+//
+//  TodoFakeModel.swift
+//  Api-ToolsTests
+//
+//  Created by Joey BARBIER on 23/07/2022.
+//
+
+import Foundation
+
+class TodoFakeModel: Decodable {
+    var id: String
+    var subtitle: String
+}