+++ Sources/Protocol/RouteProtocol.swift
@@ -19,7 +19,7 @@ public protocol Route {
 
 extension Route {
     @discardableResult
-    public func call<R: Codable> (withCompletionCall completionCall: @escaping ApiRouter.ApiResponse<R>) -> DataRequest? {
+    public func call<R: Decodable> (withCompletionCall completionCall: @escaping ApiRouter.ApiResponse<R>) -> DataRequest? {
         return ApiRouter.call(withRoute: self) { response in
             switch response.result {
             case .success(let value):
@@ -53,5 +53,34 @@ extension Route {
                 handlerResult(.failure(error))
             }
         }
+    }
+}
+
+@available(iOS 13.0.0, *)
+extension Route {
+    @discardableResult
+    public func call<R: Decodable> () async throws -> R {
+        return try await withCheckedThrowingContinuation({ (continuation: CheckedContinuation<R, Error>) in
+            self.call { (result: Result<R, Error>) in
+                continuation.resume(with: result)
+            }
+        })
+    }
+    
+    @discardableResult
+    public func call() async throws -> Data {
+        return try await withCheckedThrowingContinuation({ (continuation: CheckedContinuation<Data, Error>) in
+            self.call { (result: Result<Data, Error>) in
+                continuation.resume(with: result)
+            }
+        })
+    }
+    
+    public func call() async throws {
+        return try await withCheckedThrowingContinuation({ (continuation: CheckedContinuation<Void, Error>) in
+            self.call { (result: Result<Void, Error>) in
+                continuation.resume(with: result)
+            }
+        })
     }
 }